Issue #SB-5716 fix: Invalid qr code and new content scenario is handled by error message
