+++ src/pages/qr-code-result/qr-code-result.ts
@@ -1,3 +1,4 @@
+import { CommonUtilService } from './../../service/common-util.service';
 import { QRScannerAlert, QRAlertCallBack } from './../qrscanner/qrscanner_alert';
 import { FormAndFrameworkUtilService } from './../profile/formandframeworkutil.service';
 import {
@@ -112,7 +113,8 @@ export class QrCodeResultPage {
     private profileService: ProfileService,
     private events: Events,
     private preferences: SharedPreferences,
-    private popOverCtrl: PopoverController
+    private popOverCtrl: PopoverController,
+    private commonUtilService: CommonUtilService,
   ) {
     this.defaultImg = 'assets/imgs/ic_launcher.png';
   }
@@ -196,35 +198,9 @@ export class QrCodeResultPage {
         this.zone.run(() => {
           this.showChildrenLoader = false;
         });
-        this.showContentComingSoonAlert();
+        this.commonUtilService.showContentComingSoonAlert(this.source);
         this.navCtrl.pop();
       });
-  }
-
-  showContentComingSoonAlert() {
-    let popOver: Popover;
-    const self = this;
-    const callback: QRAlertCallBack = {
-      tryAgain() {
-        self.events.publish('event:showScanner', { pageName: PageId.LIBRARY });
-        popOver.dismiss();
-      },
-      cancel() {
-        popOver.dismiss();
-      }
-    };
-    popOver = this.popOverCtrl.create(QRScannerAlert, {
-      callback: callback,
-      icon: './assets/imgs/ic_coming_soon.png',
-      messageKey: 'CONTENT_IS_BEING_ADDED',
-      cancelKey: 'hide',
-      tryAgainKey: 'TRY_DIFF_QR',
-    }, {
-        cssClass: 'qr-alert-invalid'
-      });
-    setTimeout(() => {
-      popOver.present();
-    }, 300);
   }
 
   private showAllChild(content: any) {
