+++ src/service/common-util.service.ts
@@ -1,4 +1,5 @@
-import { LoadingController } from 'ionic-angular';
+import { QRScannerAlert } from './../pages/qrscanner/qrscanner_alert';
+import { LoadingController, Events, PopoverController } from 'ionic-angular';
 import { Injectable } from '@angular/core';
 import {
     ToastController,
@@ -7,6 +8,8 @@ import {
 import { TranslateService } from '@ngx-translate/core';
 import { Loading } from 'ionic-angular';
 import * as _ from 'lodash';
+import { Popover } from 'ionic-angular';
+import { QRAlertCallBack } from '../pages/qrscanner/qrscanner_alert';
 
 @Injectable()
 export class CommonUtilService {
@@ -14,7 +17,9 @@ export class CommonUtilService {
     constructor(
         private toastCtrl: ToastController,
         private translate: TranslateService,
-        private loadingCtrl: LoadingController
+        private loadingCtrl: LoadingController,
+        private events: Events,
+        private popOverCtrl: PopoverController
     ) {
     }
 
@@ -84,4 +89,30 @@ export class CommonUtilService {
     stringToArray(str: string = '') {
         return _.split(str, ', ');
     }
+
+    showContentComingSoonAlert(source) {
+        let popOver: Popover;
+        const self = this;
+        const callback: QRAlertCallBack = {
+          tryAgain() {
+            self.events.publish('event:showScanner', { pageName: source});
+            popOver.dismiss();
+          },
+          cancel() {
+            popOver.dismiss();
+          }
+        };
+        popOver = this.popOverCtrl.create(QRScannerAlert, {
+          callback: callback,
+          icon: './assets/imgs/ic_coming_soon.png',
+          messageKey: 'CONTENT_IS_BEING_ADDED',
+          cancelKey: 'hide',
+          tryAgainKey: 'TRY_DIFF_QR',
+        }, {
+            cssClass: 'qr-alert-invalid'
+          });
+        setTimeout(() => {
+          popOver.present();
+        }, 300);
+      }
 }
